Add resizable functionality to the code editor and preview iframe window
