+++ package.json
@@ -12,6 +12,7 @@
     "@types/prettier": "^2.7.3",
     "@types/react": "^18.2.65",
     "@types/react-dom": "^18.2.22",
+    "@types/react-resizable": "3.0.2",
     "axios": "^1.6.8",
     "bulmaswatch": "^0.8.1",
     "esbuild-wasm": "0.8.27",
@@ -20,6 +21,7 @@
     "prettier": "^2.8.8",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-resizable": "3.0.4",
     "react-scripts": "5.0.1",
     "typescript": "^4.9.5",
     "web-vitals": "^2.1.4"
