+++ src/components/code-editor.css
@@ -1,6 +1,7 @@
 .editor-wrapper {
    position: relative;
    height: 100%;
+   width: 100%;
 }
 
 .editor-wrapper .button-format {
