+++ src/components/code-editor.tsx
@@ -42,7 +42,7 @@ const CodeEditor: React.FC<CodeEditorProps> = ({ onChange, initialValue }) => {
             editorDidMount={onEditorDidMount}
             value={initialValue}
             language="javascript"
-            height="500px"
+            height="100%"
             theme='dark'
             options={{
                wordWrap: 'on',
