+++ src/components/preview.tsx
@@ -32,7 +32,7 @@ const Preview: React.FC<PreviewProps> = ({ code }) => {
       iframe.current.contentWindow.postMessage(code, '*')
    }, [code])
 
-   return <iframe title='preview' ref={iframe} sandbox='allow-scripts' srcDoc={html} />
+   return <iframe style={{ backgroundColor: 'white' }} title='preview' ref={iframe} sandbox='allow-scripts' srcDoc={html} />
 }
 
 export default Preview;
